Convert test parser format to pytest (#1634)
